Refactored XokenService
