+++ node/src/Network/Xoken/Node/XokenService.hs
@@ -90,6 +90,7 @@ import Network.Xoken.Node.GraphDB
 import Network.Xoken.Node.P2P.BlockSync
 import Network.Xoken.Node.P2P.Common
 import Network.Xoken.Node.P2P.Types
+import Network.Xoken.Node.Service
 import Network.Xoken.Util (bsToInteger, integerToBS)
 import Numeric (showHex)
 import System.Logger as LG
@@ -111,1290 +112,6 @@ data EndPointConnection =
         , encodingFormat :: IORef EncodingFormat
         }
 
-xGetChainInfo :: (HasXokenNodeEnv env m, HasLogger m, MonadIO m) => m (Maybe ChainInfo)
-xGetChainInfo = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB $ dbe
-        str = "SELECT key,value from xoken.misc_store"
-        qstr = str :: Q.QueryString Q.R () (DT.Text, (Maybe Bool, Int32, Maybe Int64, DT.Text))
-        p = Q.defQueryParams Q.One ()
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if L.length iop < 3
-                then do
-                    return Nothing
-                else do
-                    let (_, blocks, _, bestSyncedHash) = snd . head $ (L.filter (\x -> fst x == "best-synced") iop)
-                        (_, headers, _, bestBlockHash) = snd . head $ (L.filter (\x -> fst x == "best_chain_tip") iop)
-                        (_, lagHeight, _, chainwork) = snd . head $ (L.filter (\x -> fst x == "chain-work") iop)
-                    blk <- xGetBlockHeight headers
-                    lagCW <- calculateChainWork [(lagHeight + 1) .. (headers)] conn
-                    case blk of
-                        Nothing -> return Nothing
-                        Just b -> do
-                            return $
-                                Just $
-                                ChainInfo
-                                    "main"
-                                    (showHex (lagCW + (read . DT.unpack $ chainwork)) "")
-                                    (convertBitsToDifficulty . blockBits' . rbHeader $ b)
-                                    (headers)
-                                    (blocks)
-                                    (DT.unpack bestBlockHash)
-                                    (DT.unpack bestSyncedHash)
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetChainInfo: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetChainHeaders :: (HasXokenNodeEnv env m, HasLogger m, MonadIO m) => Int32 -> Int -> m [ChainHeader]
-xGetChainHeaders sblk pgsize = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        str = "SELECT block_hash,block_height,tx_count,block_header from xoken.blocks_by_height where block_height in ?"
-        qstr = str :: Q.QueryString Q.R (Identity [Int32]) (DT.Text, Int32, Maybe Int32, DT.Text)
-        p = Q.defQueryParams Q.One $ Identity (L.take pgsize [sblk ..])
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return []
-                else do
-                    case traverse
-                             (\(hash, ht, txc, hdr) ->
-                                  case (eitherDecode $ BSL.fromStrict $ DTE.encodeUtf8 hdr) of
-                                      (Right bh) ->
-                                          Right $ ChainHeader ht (DT.unpack hash) bh (maybe (-1) fromIntegral txc)
-                                      Left e -> Left e)
-                             (iop) of
-                        Right x -> return x
-                        Left e -> do
-                            err lg $ LG.msg $ "Error: xGetChainHeaders: decode failed for blockrecord: " <> show e
-                            return []
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetChainHeaders: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetBlockHash :: (HasXokenNodeEnv env m, MonadIO m) => DT.Text -> m (Maybe BlockRecord)
-xGetBlockHash hash = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        str =
-            "SELECT block_hash,block_height,block_header,next_block_hash,block_size,tx_count,coinbase_tx from xoken.blocks_by_hash where block_hash = ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity DT.Text) ( DT.Text
-                                                        , Int32
-                                                        , DT.Text
-                                                        , Maybe DT.Text
-                                                        , Maybe Int32
-                                                        , Maybe Int32
-                                                        , Maybe Blob)
-        p = Q.defQueryParams Q.One $ Identity hash
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return Nothing
-                else do
-                    let (hs, ht, hdr, nbhs, size, txc, cbase) = iop !! 0
-                    case eitherDecode $ BSL.fromStrict $ DTE.encodeUtf8 hdr of
-                        Right bh ->
-                            return $
-                            Just $
-                            BlockRecord
-                                (fromIntegral ht)
-                                (DT.unpack hs)
-                                bh
-                                (maybe "" DT.unpack nbhs)
-                                (maybe (-1) fromIntegral size)
-                                (maybe (-1) fromIntegral txc)
-                                ("")
-                                (maybe "" (coinbaseTxToMessage . fromBlob) cbase)
-                                (maybe "" fromBlob cbase)
-                        Left err -> do
-                            liftIO $ print $ "Decode failed with error: " <> show err
-                            return Nothing
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetBlocksHash: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetBlocksHashes :: (HasXokenNodeEnv env m, MonadIO m) => [DT.Text] -> m ([BlockRecord])
-xGetBlocksHashes hashes = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        str =
-            "SELECT block_hash,block_height,block_header,next_block_hash,block_size,tx_count,coinbase_tx from xoken.blocks_by_hash where block_hash in ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity [DT.Text]) ( DT.Text
-                                                          , Int32
-                                                          , DT.Text
-                                                          , Maybe DT.Text
-                                                          , Maybe Int32
-                                                          , Maybe Int32
-                                                          , Maybe Blob)
-        p = Q.defQueryParams Q.One $ Identity $ hashes
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return []
-                else do
-                    case traverse
-                             (\(hs, ht, hdr, nbhs, size, txc, cbase) ->
-                                  case (eitherDecode $ BSL.fromStrict $ DTE.encodeUtf8 hdr) of
-                                      (Right bh) ->
-                                          Right $
-                                          BlockRecord
-                                              (fromIntegral ht)
-                                              (DT.unpack hs)
-                                              bh
-                                              (maybe "" DT.unpack nbhs)
-                                              (maybe (-1) fromIntegral size)
-                                              (maybe (-1) fromIntegral txc)
-                                              ("")
-                                              (maybe "" (coinbaseTxToMessage . fromBlob) cbase)
-                                              (maybe "" fromBlob cbase)
-                                      Left err -> Left err)
-                             (iop) of
-                        Right x -> return x
-                        Left err -> do
-                            liftIO $ print $ "decode failed for blockrecord: " <> show err
-                            return []
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetBlocksHashes: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetBlockHeight :: (HasXokenNodeEnv env m, MonadIO m) => Int32 -> m (Maybe BlockRecord)
-xGetBlockHeight height = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        str =
-            "SELECT block_hash,block_height,block_header,next_block_hash,block_size,tx_count,coinbase_tx from xoken.blocks_by_height where block_height = ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity Int32) ( DT.Text
-                                                      , Int32
-                                                      , DT.Text
-                                                      , Maybe DT.Text
-                                                      , Maybe Int32
-                                                      , Maybe Int32
-                                                      , Maybe Blob)
-        p = Q.defQueryParams Q.One $ Identity height
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return Nothing
-                else do
-                    let (hs, ht, hdr, nbhs, size, txc, cbase) = iop !! 0
-                    case eitherDecode $ BSL.fromStrict $ DTE.encodeUtf8 hdr of
-                        Right bh -> do
-                            return $
-                                Just $
-                                BlockRecord
-                                    (fromIntegral ht)
-                                    (DT.unpack hs)
-                                    bh
-                                    (maybe "" DT.unpack nbhs)
-                                    (maybe (-1) fromIntegral size)
-                                    (maybe (-1) fromIntegral txc)
-                                    ("")
-                                    (maybe "" (coinbaseTxToMessage . fromBlob) cbase)
-                                    (maybe "" fromBlob cbase)
-                        Left err -> do
-                            liftIO $ print $ "Decode failed with error: " <> show err
-                            return Nothing
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetBlockHeight: " <> show e
-            throw KeyValueDBLookupException
-
-xGetTxOutputSpendStatus :: (HasXokenNodeEnv env m, MonadIO m) => String -> Int32 -> m (Maybe TxOutputSpendStatus)
-xGetTxOutputSpendStatus txId outputIndex = do
-    dbe <- getDB
-    let conn = keyValDB (dbe)
-        str = "SELECT is_recv, block_info, other FROM xoken.txid_outputs WHERE txid=? AND output_index=?"
-        qstr =
-            str :: Q.QueryString Q.R (DT.Text, Int32) ( Bool
-                                                      , (DT.Text, Int32, Int32)
-                                                      , Set ((DT.Text, Int32), Int32, (DT.Text, Int64)))
-        p = Q.defQueryParams Q.One (DT.pack txId, outputIndex)
-    iop <- Q.runClient conn (Q.query qstr p)
-    if length iop == 0
-        then return Nothing
-        else do
-            if L.length iop == 1
-                then return $ Just $ TxOutputSpendStatus False Nothing Nothing Nothing
-                else do
-                    let siop = L.sortBy (\(x, _, _) (y, _, _) -> compare x y) iop
-                        (_, (_, spendingTxBlkHeight, _), other) = siop !! 0
-                        ((spendingTxID, _), spendingTxIndex, _) = head $ DCP.fromSet other
-                    return $
-                        Just $
-                        TxOutputSpendStatus
-                            True
-                            (Just $ DT.unpack spendingTxID)
-                            (Just spendingTxBlkHeight)
-                            (Just spendingTxIndex)
-
-xGetBlocksHeights :: (HasXokenNodeEnv env m, MonadIO m) => [Int32] -> m ([BlockRecord])
-xGetBlocksHeights heights = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        str =
-            "SELECT block_hash,block_height,block_header,next_block_hash,block_size,tx_count,coinbase_tx from xoken.blocks_by_height where block_height in ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity [Int32]) ( DT.Text
-                                                        , Int32
-                                                        , DT.Text
-                                                        , Maybe DT.Text
-                                                        , Maybe Int32
-                                                        , Maybe Int32
-                                                        , Maybe Blob)
-        p = Q.defQueryParams Q.One $ Identity $ heights
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return []
-                else do
-                    case traverse
-                             (\(hs, ht, hdr, nbhs, size, txc, cbase) ->
-                                  case (eitherDecode $ BSL.fromStrict $ DTE.encodeUtf8 hdr) of
-                                      (Right bh) ->
-                                          Right $
-                                          BlockRecord
-                                              (fromIntegral ht)
-                                              (DT.unpack hs)
-                                              bh
-                                              (maybe "" DT.unpack nbhs)
-                                              (maybe (-1) fromIntegral size)
-                                              (maybe (-1) fromIntegral txc)
-                                              ("")
-                                              (maybe "" (coinbaseTxToMessage . fromBlob) cbase)
-                                              (maybe "" fromBlob cbase)
-                                      Left err -> Left err)
-                             (iop) of
-                        Right x -> return x
-                        Left err -> do
-                            liftIO $ print $ "decode failed for blockrecord: " <> show err
-                            return []
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetBlockHeights: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetTxIDsByBlockHash :: (HasXokenNodeEnv env m, HasLogger m, MonadIO m) => String -> Int32 -> Int32 -> m [String]
-xGetTxIDsByBlockHash hash pgSize pgNum = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB $ dbe
-        txsToSkip = pgSize * (pgNum - 1)
-        firstPage = (+ 1) $ fromIntegral $ floor $ (fromIntegral txsToSkip) / 100
-        lastPage = (+ 1) $ fromIntegral $ floor $ (fromIntegral $ txsToSkip + pgSize) / 100
-        txDropFromFirst = fromIntegral $ txsToSkip `mod` 100
-        str = "SELECT page_number, txids from xoken.blockhash_txids where block_hash = ? and page_number in ? "
-        qstr = str :: Q.QueryString Q.R (DT.Text, [Int32]) (Int32, [DT.Text])
-        p = Q.defQueryParams Q.One $ (DT.pack hash, [firstPage .. lastPage])
-    res <- liftIO $ try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop ->
-            return . L.take (fromIntegral pgSize) . L.drop txDropFromFirst . L.concat $ (fmap DT.unpack . snd) <$> iop
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetTxIDsByBlockHash: " <> show e
-            throw KeyValueDBLookupException
-
-xGetTxHash :: (HasXokenNodeEnv env m, MonadIO m) => DT.Text -> m (Maybe RawTxRecord)
-xGetTxHash hash = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        net = NC.bitcoinNetwork $ nodeConfig bp2pEnv
-        str = "SELECT tx_id, block_info, tx_serialized, inputs, fees from xoken.transactions where tx_id = ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity DT.Text) ( DT.Text
-                                                        , (DT.Text, Int32, Int32)
-                                                        , Blob
-                                                        , Set ((DT.Text, Int32), Int32, (DT.Text, Int64))
-                                                        , Int64)
-        p = Q.defQueryParams Q.One $ Identity $ hash
-    res <-
-        LE.try $
-        LA.concurrently
-            (LA.concurrently (Q.runClient conn (Q.query qstr p)) (getTxOutputsFromTxId hash))
-            (xGetMerkleBranch $ DT.unpack hash)
-    case res of
-        Right ((iop, outs), mrkl) ->
-            if length iop == 0
-                then return Nothing
-                else do
-                    let (txid, (bhash, blkht, txind), sz, sinps, fees) = iop !! 0
-                        inps = L.sortBy (\(_, x, _) (_, y, _) -> compare x y) $ DCP.fromSet sinps
-                        tx = fromJust $ Extra.hush $ S.decodeLazy $ fromBlob sz
-                    return $
-                        Just $
-                        RawTxRecord
-                            (DT.unpack txid)
-                            (fromIntegral $ C.length $ fromBlob sz)
-                            (BlockInfo' (DT.unpack bhash) (fromIntegral blkht) (fromIntegral txind))
-                            (fromBlob sz)
-                            (zipWith mergeTxOutTxOutput (txOut tx) outs)
-                            (zipWith mergeTxInTxInput (txIn tx) $
-                             (\((outTxId, outTxIndex), inpTxIndex, (addr, value)) ->
-                                  TxInput (DT.unpack outTxId) outTxIndex inpTxIndex (DT.unpack addr) value "") <$>
-                             inps)
-                            fees
-                            mrkl
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetTxHash: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetTxHashes :: (HasXokenNodeEnv env m, MonadIO m) => [DT.Text] -> m ([RawTxRecord])
-xGetTxHashes hashes = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        net = NC.bitcoinNetwork $ nodeConfig bp2pEnv
-        str = "SELECT tx_id, block_info, tx_serialized, inputs, fees from xoken.transactions where tx_id in ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity [DT.Text]) ( DT.Text
-                                                          , (DT.Text, Int32, Int32)
-                                                          , Blob
-                                                          , Set ((DT.Text, Int32), Int32, (DT.Text, Int64))
-                                                          , Int64)
-        p = Q.defQueryParams Q.One $ Identity $ hashes
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            txRecs <-
-                traverse
-                    (\(txid, (bhash, blkht, txind), sz, sinps, fees) -> do
-                         let inps = L.sortBy (\(_, x, _) (_, y, _) -> compare x y) $ DCP.fromSet sinps
-                             tx = fromJust $ Extra.hush $ S.decodeLazy $ fromBlob sz
-                         res' <-
-                             LE.try $ LA.concurrently (getTxOutputsFromTxId txid) (xGetMerkleBranch $ DT.unpack txid)
-                         case res' of
-                             Right (outs, mrkl) ->
-                                 return $
-                                 Just $
-                                 RawTxRecord
-                                     (DT.unpack txid)
-                                     (fromIntegral $ C.length $ fromBlob sz)
-                                     (BlockInfo' (DT.unpack bhash) (fromIntegral blkht) (fromIntegral txind))
-                                     (fromBlob sz)
-                                     (zipWith mergeTxOutTxOutput (txOut tx) outs)
-                                     (zipWith mergeTxInTxInput (txIn tx) $
-                                      (\((outTxId, outTxIndex), inpTxIndex, (addr, value)) ->
-                                           TxInput (DT.unpack outTxId) outTxIndex inpTxIndex (DT.unpack addr) value "") <$>
-                                      inps)
-                                     fees
-                                     mrkl
-                             Left (e :: SomeException) -> do
-                                 err lg $ LG.msg $ "Error: xGetTxHashes: " ++ show e
-                                 return Nothing)
-                    iop
-            return $ fromMaybe [] (sequence txRecs)
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetTxHashes: " ++ show e
-            throw KeyValueDBLookupException
-
-getTxOutputsFromTxId :: (HasXokenNodeEnv env m, HasLogger m, MonadIO m) => DT.Text -> m [TxOutput]
-getTxOutputsFromTxId txid = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        toStr = "SELECT output_index,block_info,is_recv,other,value,address FROM xoken.txid_outputs WHERE txid=?"
-        toQStr =
-            toStr :: Q.QueryString Q.R (Identity DT.Text) ( Int32
-                                                          , (DT.Text, Int32, Int32)
-                                                          , Bool
-                                                          , Set ((DT.Text, Int32), Int32, (DT.Text, Int64))
-                                                          , Int64
-                                                          , DT.Text)
-        par = Q.defQueryParams Q.One (Identity txid)
-    res <- LE.try $ Q.runClient conn (Q.query toQStr par)
-    case res of
-        Right t -> do
-            if length t == 0
-                then do
-                    err lg $ LG.msg $ "Error: getTxOutputsFromTxId: No entry in txid_outputs for txid: " ++ show txid
-                    return []
-                else do
-                    let txg =
-                            (L.sortBy (\(_, _, x, _, _, _) (_, _, y, _, _, _) -> compare x y)) <$>
-                            (L.groupBy (\(x, _, _, _, _, _) (y, _, _, _, _, _) -> x == y) t)
-                        txOutData =
-                            (\inp ->
-                                 case inp of
-                                     [(idx, bif, recv, oth, val, addr)] ->
-                                         genTxOutputData (txid, idx, (bif, recv, oth, val, addr), Nothing)
-                                     [(idx1, bif1, recv1, oth1, val1, addr1), (_, bif2, recv2, oth2, val2, addr2)] ->
-                                         genTxOutputData
-                                             ( txid
-                                             , idx1
-                                             , (bif2, recv2, oth2, val2, addr2)
-                                             , Just (bif1, recv1, oth1, val1, addr1))) <$>
-                            txg
-                    return $ txOutputDataToOutput <$> txOutData
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: getTxOutputsFromTxId: " ++ show e
-            throw KeyValueDBLookupException
-
-getTxOutputsData :: (HasXokenNodeEnv env m, HasLogger m, MonadIO m) => (DT.Text, Int32) -> m TxOutputData
-getTxOutputsData (txid, index) = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        toStr = "SELECT block_info,is_recv,other,value,address FROM xoken.txid_outputs WHERE txid=? AND output_index=?"
-        toQStr =
-            toStr :: Q.QueryString Q.R (DT.Text, Int32) ( (DT.Text, Int32, Int32)
-                                                        , Bool
-                                                        , Set ((DT.Text, Int32), Int32, (DT.Text, Int64))
-                                                        , Int64
-                                                        , DT.Text)
-        top = Q.defQueryParams Q.One (txid, index)
-    toRes <- LE.try $ Q.runClient conn (Q.query toQStr top)
-    case toRes of
-        Right es -> do
-            if length es == 0
-                then do
-                    err lg $
-                        LG.msg $
-                        "Error: getTxOutputsData: No entry in txid_outputs for (txid,index): " ++ show (txid, index)
-                    throw KeyValueDBLookupException
-                else do
-                    let txg = L.sortBy (\(_, x, _, _, _) (_, y, _, _, _) -> compare x y) es
-                    return $
-                        case txg of
-                            [x] -> genTxOutputData (txid, index, x, Nothing)
-                            [x, y] -> genTxOutputData (txid, index, y, Just x)
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: getTxOutputsData: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetOutputsAddress ::
-       (HasXokenNodeEnv env m, MonadIO m)
-    => String
-    -> Maybe Int32
-    -> Maybe Int64
-    -> m ([ResultWithCursor AddressOutputs Int64])
-xGetOutputsAddress address pgSize mbNomTxInd = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        net = NC.bitcoinNetwork $ nodeConfig bp2pEnv
-        nominalTxIndex =
-            case mbNomTxInd of
-                (Just n) -> n
-                Nothing -> maxBound
-        sh = convertToScriptHash net address
-        str = "SELECT nominal_tx_index,output FROM xoken.script_hash_outputs WHERE script_hash=? AND nominal_tx_index<?"
-        qstr = str :: Q.QueryString Q.R (DT.Text, Int64) (Int64, (DT.Text, Int32))
-        aop = Q.defQueryParams Q.One (DT.pack address, nominalTxIndex)
-        shp = Q.defQueryParams Q.One (maybe "" DT.pack sh, nominalTxIndex)
-    res <-
-        LE.try $
-        LA.concurrently
-            (case sh of
-                 Nothing -> return []
-                 Just s -> Q.runClient conn (Q.query qstr (shp {pageSize = pgSize})))
-            (case address of
-                 ('3':_) -> return []
-                 _ -> Q.runClient conn (Q.query qstr (aop {pageSize = pgSize})))
-    case res of
-        Right (sr, ar) -> do
-            let iops =
-                    fmap head $
-                    L.groupBy (\(x, _) (y, _) -> x == y) $
-                    L.sortBy
-                        (\(x, _) (y, _) ->
-                             if x < y
-                                 then GT
-                                 else LT)
-                        (sr ++ ar)
-                iop =
-                    case pgSize of
-                        Nothing -> iops
-                        (Just pg) -> L.take (fromIntegral pg) iops
-            if length iop == 0
-                then return []
-                else do
-                    res' <- sequence $ (\(_, (txid, index)) -> getTxOutputsData (txid, index)) <$> iop
-                    return $
-                        ((\((nti, (op_txid, op_txidx)), TxOutputData _ _ _ val bi ips si) ->
-                              ResultWithCursor
-                                  (AddressOutputs
-                                       (address)
-                                       (OutPoint' (DT.unpack op_txid) (fromIntegral op_txidx))
-                                       bi
-                                       si
-                                       ((\((oph, opi), ii, (_, ov)) ->
-                                             ( OutPoint' (DT.unpack oph) (fromIntegral opi)
-                                             , fromIntegral ii
-                                             , fromIntegral ov)) <$>
-                                        ips)
-                                       val)
-                                  nti) <$>)
-                            (zip iop res')
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetOutputsAddress':" ++ show e
-            throw KeyValueDBLookupException
-
-xGetUTXOsAddress ::
-       (HasXokenNodeEnv env m, MonadIO m)
-    => String
-    -> Maybe Int32
-    -> Maybe (DT.Text, Int32)
-    -> m ([ResultWithCursor AddressOutputs (DT.Text, Int32)])
-xGetUTXOsAddress address pgSize mbFromOutput = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        net = NC.bitcoinNetwork $ nodeConfig bp2pEnv
-        sh = convertToScriptHash net address
-        fromOutput =
-            case mbFromOutput of
-                (Just n) -> n
-                Nothing -> maxBoundOutput
-        str = "SELECT output FROM xoken.script_hash_unspent_outputs WHERE script_hash=? AND output<?"
-        qstr = str :: Q.QueryString Q.R (DT.Text, (DT.Text, Int32)) (Identity (DT.Text, Int32))
-        aop = Q.defQueryParams Q.One (DT.pack address, fromOutput)
-        shp = Q.defQueryParams Q.One (maybe "" DT.pack sh, fromOutput)
-    res <-
-        LE.try $
-        LA.concurrently
-            (case sh of
-                 Nothing -> return []
-                 Just s -> Q.runClient conn (Q.query qstr (shp {pageSize = pgSize})))
-            (case address of
-                 ('3':_) -> return []
-                 _ -> Q.runClient conn (Q.query qstr (aop {pageSize = pgSize})))
-    case res of
-        Right (sr, ar) -> do
-            let iops =
-                    fmap head $
-                    L.groupBy (\(Identity x) (Identity y) -> x == y) $
-                    L.sortBy
-                        (\(Identity x) (Identity y) ->
-                             if x < y
-                                 then GT
-                                 else LT)
-                        (sr ++ ar)
-                iop =
-                    case pgSize of
-                        Nothing -> iops
-                        (Just pg) -> L.take (fromIntegral pg) iops
-            if length iop == 0
-                then return []
-                else do
-                    res' <- sequence $ (\(Identity (txid, index)) -> getTxOutputsData (txid, index)) <$> iop
-                    return $
-                        ((\((Identity (op_txid, op_txidx)), TxOutputData _ _ _ val bi ips si) ->
-                              ResultWithCursor
-                                  (AddressOutputs
-                                       (address)
-                                       (OutPoint' (DT.unpack op_txid) (fromIntegral op_txidx))
-                                       bi
-                                       si
-                                       ((\((oph, opi), ii, (_, ov)) ->
-                                             ( OutPoint' (DT.unpack oph) (fromIntegral opi)
-                                             , fromIntegral ii
-                                             , fromIntegral ov)) <$>
-                                        ips)
-                                       val)
-                                  (op_txid, op_txidx)) <$>)
-                            (zip iop res')
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetUTXOsAddress:" ++ show e
-            throw KeyValueDBLookupException
-
-xGetOutputsScriptHash ::
-       (HasXokenNodeEnv env m, MonadIO m)
-    => String
-    -> Maybe Int32
-    -> Maybe Int64
-    -> m ([ResultWithCursor ScriptOutputs Int64])
-xGetOutputsScriptHash scriptHash pgSize mbNomTxInd = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        nominalTxIndex =
-            case mbNomTxInd of
-                (Just n) -> n
-                Nothing -> maxBound
-        str =
-            "SELECT script_hash,nominal_tx_index,output FROM xoken.script_hash_outputs WHERE script_hash=? AND nominal_tx_index<?"
-        qstr = str :: Q.QueryString Q.R (DT.Text, Int64) (DT.Text, Int64, (DT.Text, Int32))
-        par = Q.defQueryParams Q.One (DT.pack scriptHash, nominalTxIndex)
-    res <- LE.try $ Q.runClient conn (Q.query qstr (par {pageSize = pgSize}))
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return []
-                else do
-                    res <- sequence $ (\(_, _, (txid, index)) -> getTxOutputsData (txid, index)) <$> iop
-                    return $
-                        ((\((addr, nti, (op_txid, op_txidx)), TxOutputData _ _ _ val bi ips si) ->
-                              ResultWithCursor
-                                  (ScriptOutputs
-                                       (DT.unpack addr)
-                                       (OutPoint' (DT.unpack op_txid) (fromIntegral op_txidx))
-                                       bi
-                                       si
-                                       ((\((oph, opi), ii, (_, ov)) ->
-                                             ( OutPoint' (DT.unpack oph) (fromIntegral opi)
-                                             , fromIntegral ii
-                                             , fromIntegral ov)) <$>
-                                        ips)
-                                       val)
-                                  nti) <$>)
-                            (zip iop res)
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetOutputsScriptHash':" ++ show e
-            throw KeyValueDBLookupException
-
-xGetUTXOsScriptHash ::
-       (HasXokenNodeEnv env m, MonadIO m)
-    => String
-    -> Maybe Int32
-    -> Maybe (DT.Text, Int32)
-    -> m ([ResultWithCursor ScriptOutputs (DT.Text, Int32)])
-xGetUTXOsScriptHash scriptHash pgSize mbFromOutput = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB (dbe)
-        fromOutput =
-            case mbFromOutput of
-                (Just n) -> n
-                Nothing -> maxBoundOutput
-        str = "SELECT script_hash,output FROM xoken.script_hash_unspent_outputs WHERE script_hash=? AND output<?"
-        qstr = str :: Q.QueryString Q.R (DT.Text, (DT.Text, Int32)) (DT.Text, (DT.Text, Int32))
-        par = Q.defQueryParams Q.One (DT.pack scriptHash, fromOutput)
-    res <- LE.try $ Q.runClient conn (Q.query qstr (par {pageSize = pgSize}))
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return []
-                else do
-                    res <- sequence $ (\(_, (txid, index)) -> getTxOutputsData (txid, index)) <$> iop
-                    return $
-                        ((\((addr, (op_txid, op_txidx)), TxOutputData _ _ _ val bi ips si) ->
-                              ResultWithCursor
-                                  (ScriptOutputs
-                                       (DT.unpack addr)
-                                       (OutPoint' (DT.unpack op_txid) (fromIntegral op_txidx))
-                                       bi
-                                       si
-                                       ((\((oph, opi), ii, (_, ov)) ->
-                                             ( OutPoint' (DT.unpack oph) (fromIntegral opi)
-                                             , fromIntegral ii
-                                             , fromIntegral ov)) <$>
-                                        ips)
-                                       val)
-                                  (op_txid, op_txidx)) <$>)
-                            (zip iop res)
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetUTXOsScriptHash':" ++ show e
-            throw KeyValueDBLookupException
-
-runWithManyInputs ::
-       (HasXokenNodeEnv env m, MonadIO m, Ord c, Eq r, Integral p, Bounded p)
-    => (i -> Maybe p -> Maybe c -> m ([ResultWithCursor r c]))
-    -> [i]
-    -> Maybe p
-    -> Maybe c
-    -> m ([ResultWithCursor r c])
-runWithManyInputs fx inputs mbPgSize cursor = do
-    let pgSize =
-            fromIntegral $
-            case mbPgSize of
-                Just ps -> ps
-                Nothing -> maxBound
-    li <- LA.mapConcurrently (\input -> fx input mbPgSize cursor) inputs
-    return $ (L.take pgSize . sort . concat $ li)
-
-xGetMerkleBranch :: (HasXokenNodeEnv env m, MonadIO m) => String -> m ([MerkleBranchNode'])
-xGetMerkleBranch txid = do
-    dbe <- getDB
-    lg <- getLogger
-    res <- liftIO $ try $ withResource (pool $ graphDB dbe) (`BT.run` queryMerkleBranch (DT.pack txid))
-    case res of
-        Right mb -> do
-            return $ Data.List.map (\x -> MerkleBranchNode' (DT.unpack $ _nodeValue x) (_isLeftNode x)) mb
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetMerkleBranch: " ++ show e
-            throw KeyValueDBLookupException
-
-xGetAllegoryNameBranch :: (HasXokenNodeEnv env m, MonadIO m) => String -> Bool -> m ([(OutPoint', [MerkleBranchNode'])])
-xGetAllegoryNameBranch name isProducer = do
-    dbe <- getDB
-    lg <- getLogger
-    res <- liftIO $ try $ withResource (pool $ graphDB dbe) (`BT.run` queryAllegoryNameBranch (DT.pack name) isProducer)
-    case res of
-        Right nb -> do
-            liftIO $
-                mapConcurrently
-                    (\x -> do
-                         let sp = DT.split (== ':') x
-                         let txid = DT.unpack $ sp !! 0
-                         let index = readMaybe (DT.unpack $ sp !! 1) :: Maybe Int32
-                         case index of
-                             Just i -> do
-                                 rs <-
-                                     liftIO $
-                                     try $ withResource (pool $ graphDB dbe) (`BT.run` queryMerkleBranch (DT.pack txid))
-                                 case rs of
-                                     Right mb -> do
-                                         let mnodes =
-                                                 Data.List.map
-                                                     (\y -> MerkleBranchNode' (DT.unpack $ _nodeValue y) (_isLeftNode y))
-                                                     mb
-                                         return $ (OutPoint' txid i, mnodes)
-                                     Left (e :: SomeException) -> do
-                                         err lg $ LG.msg $ "Error: xGetMerkleBranch: " ++ show e
-                                         throw KeyValueDBLookupException
-                             Nothing -> throw KeyValueDBLookupException)
-                    (nb)
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetAllegoryNameBranch: " ++ show e
-            throw KeyValueDBLookupException
-
-getOrMakeProducer :: (HasXokenNodeEnv env m, MonadIO m) => [Int] -> m (((OutPoint', DT.Text), Bool))
-getOrMakeProducer nameArr = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let name = DT.pack $ L.map (\x -> chr x) (nameArr)
-    let anutxos = NC.allegoryNameUtxoSatoshis $ nodeConfig $ bp2pEnv
-    res <- liftIO $ try $ withResource (pool $ graphDB dbe) (`BT.run` queryAllegoryNameScriptOp (name) True)
-    case res of
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "error fetching allegory name input :" ++ show e
-            throw e
-        Right [] -> do
-            debug lg $ LG.msg $ "allegory name not found, create recursively (1): " <> name
-            createCommitImplictTx (nameArr)
-            inres <- liftIO $ try $ withResource (pool $ graphDB dbe) (`BT.run` queryAllegoryNameScriptOp (name) True)
-            case inres of
-                Left (e :: SomeException) -> do
-                    err lg $ LG.msg $ "error fetching allegory name input :" ++ show e
-                    throw e
-                Right [] -> do
-                    err lg $ LG.msg $ "allegory name still not found, recursive create must've failed (1): " <> name
-                    throw KeyValueDBLookupException
-                Right nb -> do
-                    liftIO $ print $ "nb2~" <> show nb
-                    let sp = DT.split (== ':') $ fst (head nb)
-                    let txid = DT.unpack $ sp !! 0
-                    let index = readMaybe (DT.unpack $ sp !! 1) :: Maybe Int
-                    case index of
-                        Just i -> return $ ((OutPoint' txid (fromIntegral i), (snd $ head nb)), False)
-                        Nothing -> throw KeyValueDBLookupException
-        Right nb -> do
-            debug lg $ LG.msg $ "allegory name found! (1): " <> name
-            let sp = DT.split (== ':') $ fst (head nb)
-            let txid = DT.unpack $ sp !! 0
-            let index = readMaybe (DT.unpack $ sp !! 1) :: Maybe Int
-            case index of
-                Just i -> return $ ((OutPoint' txid (fromIntegral i), (snd $ head nb)), True)
-                Nothing -> throw KeyValueDBLookupException
-
-createCommitImplictTx :: (HasXokenNodeEnv env m, MonadIO m) => [Int] -> m ()
-createCommitImplictTx nameArr = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    alg <- getAllegory
-    let net = NC.bitcoinNetwork $ nodeConfig bp2pEnv
-    (nameip, existed) <- getOrMakeProducer (init nameArr)
-    let anutxos = NC.allegoryNameUtxoSatoshis $ nodeConfig $ bp2pEnv
-    let ins =
-            L.map
-                (\(x, s) ->
-                     TxIn (OutPoint (fromString $ opTxHash x) (fromIntegral $ opIndex x)) (fromJust $ decodeHex s) 0)
-                ([nameip])
-        -- construct OP_RETURN
-    let al =
-            Allegory
-                1
-                (init nameArr)
-                (ProducerAction
-                     (Index 0)
-                     (ProducerOutput (Index 1) (Just $ Endpoint "XokenP2P" "someuri-1"))
-                     Nothing
-                     [ (ProducerExtension
-                            (ProducerOutput (Index 2) (Just $ Endpoint "XokenP2P" "someuri-2"))
-                            (last nameArr))
-                     , (OwnerExtension (OwnerOutput (Index 3) (Just $ Endpoint "XokenP2P" "someuri-3")) (last nameArr))
-                     ])
-    let opRetScript = frameOpReturn $ C.toStrict $ serialise al
-        -- derive producer's Address
-    let prAddr = pubKeyAddr $ derivePubKeyI $ wrapSecKey True $ allegorySecretKey alg
-    let prScript = addressToScriptBS prAddr
-    let !outs = [TxOut 0 opRetScript] ++ L.map (\_ -> TxOut (fromIntegral anutxos) prScript) [1, 2, 3]
-    let !sigInputs =
-            L.map
-                (\x -> do SigInput (addressToOutput x) (fromIntegral anutxos) (prevOutput $ head ins) sigHashAll Nothing)
-                [prAddr, prAddr]
-    let psatx = Tx version ins outs locktime
-    case signTx net psatx sigInputs [allegorySecretKey alg] of
-        Right tx -> do
-            xRelayTx $ Data.Serialize.encode tx
-            return ()
-        Left err -> do
-            liftIO $ print $ "error occurred while signing the Tx: " <> show err
-            throw KeyValueDBLookupException
-  where
-    version = 1
-    locktime = 0
-
-xGetPartiallySignedAllegoryTx ::
-       (HasXokenNodeEnv env m, MonadIO m)
-    => [(OutPoint', Int)]
-    -> ([Int], Bool)
-    -> (String)
-    -> (String)
-    -> m (BC.ByteString)
-xGetPartiallySignedAllegoryTx payips (nameArr, isProducer) owner change = do
-    dbe <- getDB
-    bp2pEnv <- getBitcoinP2P
-    lg <- getLogger
-    alg <- getAllegory
-    let conn = keyValDB (dbe)
-    let net = NC.bitcoinNetwork $ nodeConfig bp2pEnv
-    -- check if name (of given type) exists
-    let name = DT.pack $ L.map (\x -> chr x) (nameArr)
-    -- read from config file
-    let anutxos = NC.allegoryNameUtxoSatoshis $ nodeConfig $ bp2pEnv
-    let feeSatsCreate = NC.allegoryTxFeeSatsProducerAction $ nodeConfig $ bp2pEnv
-    let feeSatsTransfer = NC.allegoryTxFeeSatsOwnerAction $ nodeConfig $ bp2pEnv
-    res <- liftIO $ try $ withResource (pool $ graphDB dbe) (`BT.run` queryAllegoryNameScriptOp (name) isProducer)
-    (nameip, existed) <-
-        case res of
-            Left (e :: SomeException) -> do
-                err lg $ LG.msg $ "error fetching allegory name input :" ++ show e
-                throw e
-            Right [] -> do
-                debug lg $ LG.msg $ "allegory name not found, get or make interim producers recursively : " <> name
-                getOrMakeProducer (init nameArr)
-            Right nb -> do
-                debug lg $ LG.msg $ "allegory name found! : " <> name
-                let sp = DT.split (== ':') $ fst (head nb)
-                let txid = DT.unpack $ sp !! 0
-                let index = readMaybe (DT.unpack $ sp !! 1) :: Maybe Int32
-                case index of
-                    Just i -> return $ ((OutPoint' txid i, (snd $ head nb)), True)
-                    Nothing -> throw KeyValueDBLookupException
-    inputHash <-
-        liftIO $
-        traverse
-            (\(w, _) -> do
-                 let op = OutPoint (fromString $ opTxHash w) (fromIntegral $ opIndex w)
-                 sh <- getScriptHashFromOutpoint conn (txSynchronizer bp2pEnv) lg net op 0
-                 return $ (w, ) <$> sh)
-            payips
-    let totalEffectiveInputSats = sum $ snd $ unzip payips
-    let ins =
-            L.map
-                (\(x, s) ->
-                     TxIn (OutPoint (fromString $ opTxHash x) (fromIntegral $ opIndex x)) (fromJust $ decodeHex s) 0)
-                ([nameip] ++ (catMaybes inputHash))
-    sigInputs <-
-        mapM
-            (\(x, s) -> do
-                 case (decodeOutputBS ((fst . B16.decode) (E.encodeUtf8 s))) of
-                     Left e -> do
-                         liftIO $
-                             print
-                                 ("error (allegory) unable to decode scriptOutput! | " ++
-                                  show name ++ " " ++ show (x, s) ++ " | " ++ show ((fst . B16.decode) (E.encodeUtf8 s)))
-                         throw KeyValueDBLookupException
-                     Right scr -> do
-                         return $
-                             SigInput
-                                 scr
-                                 (fromIntegral $ anutxos)
-                                 (OutPoint (fromString $ opTxHash x) (fromIntegral $ opIndex x))
-                                 sigHashAll
-                                 Nothing)
-            [nameip]
-    --
-    let outs =
-            if existed
-                then if isProducer
-                         then do
-                             let al =
-                                     Allegory
-                                         1
-                                         (init nameArr)
-                                         (ProducerAction
-                                              (Index 0)
-                                              (ProducerOutput (Index 1) (Just $ Endpoint "XokenP2P" "someuri_1"))
-                                              Nothing
-                                              [])
-                             let opRetScript = frameOpReturn $ C.toStrict $ serialise al
-                            -- derive producer's Address
-                             let prAddr = pubKeyAddr $ derivePubKeyI $ wrapSecKey True $ allegorySecretKey alg
-                             let prScript = addressToScriptBS prAddr
-                             let payAddr = pubKeyAddr $ derivePubKeyI $ wrapSecKey True $ allegorySecretKey alg
-                             let payScript = addressToScriptBS payAddr
-                             let paySats = 1000000
-                             let changeSats = totalEffectiveInputSats - (paySats + feeSatsCreate)
-                             [TxOut 0 opRetScript] ++
-                                 (L.map
-                                      (\x -> do
-                                           let addr =
-                                                   case stringToAddr net (DT.pack $ fst x) of
-                                                       Just a -> a
-                                                       Nothing -> throw InvalidOutputAddressException
-                                           let script = addressToScriptBS addr
-                                           TxOut (fromIntegral $ snd x) script)
-                                      [(owner, (fromIntegral $ anutxos)), (change, changeSats)]) ++
-                                 [TxOut ((fromIntegral paySats) :: Word64) payScript] -- the charge for the name transfer
-                         else do
-                             let al =
-                                     Allegory
-                                         1
-                                         (nameArr)
-                                         (OwnerAction
-                                              (Index 0)
-                                              (OwnerOutput (Index 1) (Just $ Endpoint "XokenP2P" "someuri_1"))
-                                              [ ProxyProvider
-                                                    "AllPay"
-                                                    "Public"
-                                                    (Endpoint "XokenP2P" "someuri_2")
-                                                    (Registration "addrCommit" "utxoCommit" "signature" 876543)
-                                              ])
-                             let opRetScript = frameOpReturn $ C.toStrict $ serialise al
-                             let payAddr = pubKeyAddr $ derivePubKeyI $ wrapSecKey True $ allegorySecretKey alg
-                             let payScript = addressToScriptBS payAddr
-                             let paySats = 1000000
-                             let changeSats = totalEffectiveInputSats - (paySats + feeSatsTransfer)
-                             [TxOut 0 opRetScript] ++
-                                 (L.map
-                                      (\x -> do
-                                           let addr =
-                                                   case stringToAddr net (DT.pack $ fst x) of
-                                                       Just a -> a
-                                                       Nothing -> throw InvalidOutputAddressException
-                                           let script = addressToScriptBS addr
-                                           TxOut (fromIntegral $ snd x) script)
-                                      [(owner, (fromIntegral $ anutxos)), (change, changeSats)]) ++
-                                 [TxOut (fromIntegral anutxos) payScript] -- the charge for the name transfer
-                else do
-                    let al =
-                            Allegory
-                                1
-                                (init nameArr)
-                                (ProducerAction
-                                     (Index 0)
-                                     (ProducerOutput (Index 1) (Just $ Endpoint "XokenP2P" "someuri_1"))
-                                     Nothing
-                                     [ OwnerExtension
-                                           (OwnerOutput (Index 2) (Just $ Endpoint "XokenP2P" "someuri_3"))
-                                           (last nameArr)
-                                     ])
-                    let opRetScript = frameOpReturn $ C.toStrict $ serialise al
-                    -- derive producer's Address
-                    let prAddr = pubKeyAddr $ derivePubKeyI $ wrapSecKey True $ allegorySecretKey alg
-                    let prScript = addressToScriptBS prAddr
-                    let payAddr = pubKeyAddr $ derivePubKeyI $ wrapSecKey True $ allegorySecretKey alg
-                    let payScript = addressToScriptBS payAddr
-                    let paySats = 1000000
-                    let changeSats = totalEffectiveInputSats - ((fromIntegral $ anutxos) + paySats + feeSatsCreate)
-                    [TxOut 0 opRetScript] ++
-                        [TxOut (fromIntegral anutxos) prScript] ++
-                        (L.map
-                             (\x -> do
-                                  let addr =
-                                          case stringToAddr net (DT.pack $ fst x) of
-                                              Just a -> a
-                                              Nothing -> throw InvalidOutputAddressException
-                                  let script = addressToScriptBS addr
-                                  TxOut (fromIntegral $ snd x) script)
-                             [(owner, (fromIntegral $ anutxos)), (change, changeSats)]) ++
-                        [TxOut ((fromIntegral paySats) :: Word64) payScript] -- the charge for the name transfer
-    --
-    let psatx = Tx version ins outs locktime
-    case signTx net psatx sigInputs [allegorySecretKey alg] of
-        Right tx -> do
-            return $ BSL.toStrict $ A.encode $ tx
-        Left err -> do
-            liftIO $ print $ "error occurred while signing the Tx: " <> show err
-            return $ BC.empty
-  where
-    version = 1
-    locktime = 0
-
-xRelayTx :: (HasXokenNodeEnv env m, MonadIO m) => BC.ByteString -> m (Bool)
-xRelayTx rawTx = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-    -- broadcast Tx
-    case runGetState (getConfirmedTx) (rawTx) 0 of
-        Left e -> do
-            err lg $ LG.msg $ "error decoding rawTx :" ++ show e
-            throw ConfirmedTxParseException
-        Right res -> do
-            debug lg $ LG.msg $ val $ "broadcasting tx"
-            case fst res of
-                Just tx -> do
-                    let outpoints = L.map (\x -> prevOutput x) (txIn tx)
-                    tr <-
-                        mapM
-                            (\x -> do
-                                 let txid = txHashToHex $ outPointHash $ prevOutput x
-                                     str =
-                                         "SELECT tx_id, block_info, tx_serialized from xoken.transactions where tx_id = ?"
-                                     qstr =
-                                         str :: Q.QueryString Q.R (Identity DT.Text) ( DT.Text
-                                                                                     , (DT.Text, Int32, Int32)
-                                                                                     , Blob)
-                                     p = Q.defQueryParams Q.One $ Identity $ (txid)
-                                 iop <- Q.runClient conn (Q.query qstr p)
-                                 if length iop == 0
-                                     then do
-                                         debug lg $ LG.msg $ "not found" ++ show txid
-                                         return Nothing
-                                     else do
-                                         let (txid, _, sz) = iop !! 0
-                                         case runGetLazy (getConfirmedTx) (fromBlob sz) of
-                                             Left e -> do
-                                                 debug lg $ LG.msg (encodeHex $ BSL.toStrict $ fromBlob sz)
-                                                 return Nothing
-                                             Right (txd) -> do
-                                                 case txd of
-                                                     Nothing -> return Nothing
-                                                     Just txn -> do
-                                                         let cout =
-                                                                 (txOut txn) !!
-                                                                 fromIntegral (outPointIndex $ prevOutput x)
-                                                         case (decodeOutputBS $ scriptOutput cout) of
-                                                             Right (so) -> do
-                                                                 return $ Just (so, outValue cout, prevOutput x)
-                                                             Left (e) -> do
-                                                                 err lg $ LG.msg $ "error decoding rawTx :" ++ show e
-                                                                 return Nothing)
-                            (txIn tx)
-                    -- if verifyStdTx net tx $ catMaybes tr
-                    --     then do
-                    allPeers <- liftIO $ readTVarIO (bitcoinPeers bp2pEnv)
-                    let !connPeers = L.filter (\x -> bpConnected (snd x)) (M.toList allPeers)
-                    debug lg $ LG.msg $ val $ "transaction verified - broadcasting tx"
-                    mapM_ (\(_, peer) -> do sendRequestMessages peer (MTx (fromJust $ fst res))) connPeers
-                    eres <- LE.try $ handleIfAllegoryTx tx True -- MUST be False
-                    case eres of
-                        Right (flg) -> return True
-                        Left (e :: SomeException) -> return False
-                Nothing -> do
-                    err lg $ LG.msg $ val $ "error decoding rawTx (2)"
-                    return $ False
-
-xGetUserByUsername :: (HasXokenNodeEnv env m, MonadIO m) => DT.Text -> m (Maybe User)
-xGetUserByUsername name = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        str =
-            "SELECT username,password,first_name,last_name,emailid,permissions,api_quota,api_used,api_expiry_time,session_key,session_key_expiry_time from xoken.user_permission where username = ?"
-        qstr =
-            str :: Q.QueryString Q.R (Identity DT.Text) ( DT.Text
-                                                        , DT.Text
-                                                        , DT.Text
-                                                        , DT.Text
-                                                        , DT.Text
-                                                        , Set DT.Text
-                                                        , Int32
-                                                        , Int32
-                                                        , UTCTime
-                                                        , DT.Text
-                                                        , UTCTime)
-        p = Q.defQueryParams Q.One $ Identity name
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return Nothing
-                else do
-                    let (uname, pwd, fname, lname, email, roles, apiQ, apiU, apiE, sk, skE) = iop !! 0
-                    userData <- liftIO $ H.lookup (userDataCache bp2pEnv) (sk)
-                    case userData of
-                        Just (_, _, used, _, _) ->
-                            return $
-                            Just $
-                            User
-                                (DT.unpack uname)
-                                (DT.unpack pwd)
-                                (DT.unpack fname)
-                                (DT.unpack lname)
-                                (DT.unpack email)
-                                (DT.unpack <$> (DCP.fromSet roles))
-                                (fromIntegral apiQ)
-                                (fromIntegral used)
-                                apiE
-                                (maskAfter 10 $ DT.unpack sk)
-                                skE
-                        Nothing ->
-                            return $
-                            Just $
-                            User
-                                (DT.unpack uname)
-                                (DT.unpack pwd)
-                                (DT.unpack fname)
-                                (DT.unpack lname)
-                                (DT.unpack email)
-                                (DT.unpack <$> (DCP.fromSet roles))
-                                (fromIntegral apiQ)
-                                (fromIntegral apiU)
-                                apiE
-                                (maskAfter 10 $ DT.unpack sk)
-                                skE
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetUserByUsername: " ++ show e
-            throw KeyValueDBLookupException
-
-xDeleteUserByUsername :: (HasXokenNodeEnv env m, MonadIO m) => DT.Text -> m ()
-xDeleteUserByUsername name = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        str = "DELETE FROM xoken.user_permission WHERE username=?"
-        qstr = str :: Q.QueryString Q.W (Identity DT.Text) ()
-        par = Q.defQueryParams Q.One (Identity name)
-    res <- liftIO $ try $ Q.runClient conn (Q.write qstr par)
-    case res of
-        Right () -> do
-            cacheList <- liftIO $ (H.toList $ userDataCache bp2pEnv)
-            liftIO $
-                mapM_
-                    (\(k, (n, _, _, _, _)) ->
-                         if n == name
-                             then H.delete (userDataCache bp2pEnv) k
-                             else return ())
-                    cacheList
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xDeleteUserByUsername: " ++ show e
-            throw e
-
-xUpdateUserByUsername :: (HasXokenNodeEnv env m, MonadIO m) => DT.Text -> UpdateUserByUsername' -> m Bool
-xUpdateUserByUsername name (UpdateUserByUsername' {..}) = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    res <- LE.try $ xGetUserByUsername name
-    case res of
-        Right (Just (User {..})) -> do
-            let conn = keyValDB (dbe)
-                str =
-                    "UPDATE xoken.user_permission SET password=?,first_name=?,last_name=?,emailid=?,api_quota=?,permissions=?,api_expiry_time=? WHERE username=?"
-                qstr =
-                    str :: Q.QueryString Q.W (DT.Text, DT.Text, DT.Text, DT.Text, Int32, Set DT.Text, UTCTime, DT.Text) ()
-                par =
-                    Q.defQueryParams
-                        Q.One
-                        ( fromMaybe
-                              (DT.pack uHashedPassword)
-                              ((encodeHex . S.encode . sha256 . B64.encode . BC.pack) <$> uuPassword)
-                        , DT.pack $ fromMaybe uFirstName uuFirstName
-                        , DT.pack $ fromMaybe uLastName uuLastName
-                        , DT.pack $ fromMaybe uEmail uuEmail
-                        , fromMaybe (fromIntegral uApiQuota) uuApiQuota
-                        , DCP.Set $ fmap DT.pack $ fromMaybe uRoles uuRoles
-                        , fromMaybe uApiExpiryTime uuApiExpiryTime
-                        , name)
-            res' <- liftIO $ try $ Q.runClient conn (Q.write qstr par)
-            case res' of
-                Right () -> do
-                    liftIO $
-                        H.mutate
-                            (userDataCache bp2pEnv)
-                            (DT.pack uSessionKey)
-                            (\v ->
-                                 ( (\(n, _, u, e, r) -> (n, fromMaybe (fromIntegral uApiQuota) uuApiQuota, u, e, r)) <$>
-                                   v
-                                 , True))
-                Left (e :: SomeException) -> do
-                    err lg $ LG.msg $ "Error: xUpdateUserByUsername: " ++ show e
-                    throw e
-        Right Nothing -> return False
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xUpdateUserByUsername: " ++ show e
-            throw e
-
-xGetUserBySessionKey :: (HasXokenNodeEnv env m, MonadIO m) => DT.Text -> m (Maybe User)
-xGetUserBySessionKey skey = do
-    dbe <- getDB
-    lg <- getLogger
-    bp2pEnv <- getBitcoinP2P
-    let conn = keyValDB (dbe)
-        str =
-            "SELECT username,first_name,last_name,emailid,permissions,api_quota,api_used,api_expiry_time,session_key,session_key_expiry_time from xoken.user_permission where session_key = ? ALLOW FILTERING "
-        qstr =
-            str :: Q.QueryString Q.R (Identity DT.Text) ( DT.Text
-                                                        , DT.Text
-                                                        , DT.Text
-                                                        , DT.Text
-                                                        , Set DT.Text
-                                                        , Int32
-                                                        , Int32
-                                                        , UTCTime
-                                                        , DT.Text
-                                                        , UTCTime)
-        p = Q.defQueryParams Q.One $ Identity skey
-    res <- LE.try $ Q.runClient conn (Q.query qstr p)
-    case res of
-        Right iop -> do
-            if length iop == 0
-                then return Nothing
-                else do
-                    let (uname, fname, lname, email, roles, apiQ, apiU, apiE, sk, skE) = iop !! 0
-                    userData <- liftIO $ H.lookup (userDataCache bp2pEnv) (sk)
-                    case userData of
-                        Just (_, _, used, _, _) ->
-                            return $
-                            Just $
-                            User
-                                (DT.unpack uname)
-                                (DT.unpack "")
-                                (DT.unpack fname)
-                                (DT.unpack lname)
-                                (DT.unpack email)
-                                (DT.unpack <$> (DCP.fromSet roles))
-                                (fromIntegral apiQ)
-                                (fromIntegral used)
-                                apiE
-                                (maskAfter 10 $ DT.unpack sk)
-                                skE
-                        Nothing ->
-                            return $
-                            Just $
-                            User
-                                (DT.unpack uname)
-                                (DT.unpack "")
-                                (DT.unpack fname)
-                                (DT.unpack lname)
-                                (DT.unpack email)
-                                (DT.unpack <$> (DCP.fromSet roles))
-                                (fromIntegral apiQ)
-                                (fromIntegral apiU)
-                                apiE
-                                (maskAfter 10 $ DT.unpack sk)
-                                skE
-        Left (e :: SomeException) -> do
-            err lg $ LG.msg $ "Error: xGetUserBySessionKey: " ++ show e
-            throw KeyValueDBLookupException
-
 authLoginClient ::
        (HasXokenNodeEnv env m, MonadIO m) => RPCMessage -> Network -> EndPointConnection -> Bool -> m (RPCMessage)
 authLoginClient msg net epConn pretty = do
@@ -1408,47 +125,6 @@ authLoginClient msg net epConn pretty = do
                     return $ RPCResponse 200 pretty $ Right $ Just $ AuthenticateResp resp
                 ___ -> return $ RPCResponse 404 pretty $ Left $ RPCError INVALID_REQUEST Nothing
         _____ -> return $ RPCResponse 200 pretty $ Right $ Just $ AuthenticateResp $ AuthResp Nothing 0 0
-
-login :: (MonadIO m, HasXokenNodeEnv env m) => DT.Text -> BC.ByteString -> m AuthResp
-login user pass = do
-    dbe <- getDB
-    lg <- getLogger
-    let conn = keyValDB dbe
-        hashedPasswd = encodeHex ((S.encode $ sha256 pass))
-        str =
-            " SELECT password, api_quota, api_used, session_key_expiry_time FROM xoken.user_permission WHERE username = ? "
-        qstr = str :: Q.QueryString Q.R (Identity DT.Text) (DT.Text, Int32, Int32, UTCTime)
-        p = Q.defQueryParams Q.One $ Identity $ user
-    res <- liftIO $ try $ Q.runClient conn (Q.query (Q.prepared qstr) p)
-    case res of
-        Left (SomeException e) -> do
-            err lg $ LG.msg $ "Error: SELECT'ing from 'user_permission': " ++ show e
-            throw e
-        Right (op) -> do
-            if length op == 0
-                then return $ AuthResp Nothing 0 0
-                else do
-                    case (op !! 0) of
-                        (sk, _, _, _) -> do
-                            if (sk /= hashedPasswd)
-                                then return $ AuthResp Nothing 0 0
-                                else do
-                                    tm <- liftIO $ getCurrentTime
-                                    newSessionKey <- liftIO $ generateSessionKey
-                                    let str1 =
-                                            "UPDATE xoken.user_permission SET session_key = ?, session_key_expiry_time = ? WHERE username = ? "
-                                        qstr1 = str1 :: Q.QueryString Q.W (DT.Text, UTCTime, DT.Text) ()
-                                        par1 =
-                                            Q.defQueryParams
-                                                Q.One
-                                                (newSessionKey, (addUTCTime (nominalDay * 30) tm), user)
-                                    res1 <- liftIO $ try $ Q.runClient conn (Q.write (qstr1) par1)
-                                    case res1 of
-                                        Right () -> return ()
-                                        Left (SomeException e) -> do
-                                            err lg $ LG.msg $ "Error: UPDATE'ing into 'user_permission': " ++ show e
-                                            throw e
-                                    return $ AuthResp (Just $ DT.unpack newSessionKey) 1 100
 
 delegateRequest :: (HasXokenNodeEnv env m, MonadIO m) => RPCMessage -> EndPointConnection -> Network -> m (RPCMessage)
 delegateRequest encReq epConn net = do
@@ -1806,37 +482,3 @@ goGetResource msg net roles sessKey pretty = do
                     return $ RPCResponse 200 pretty $ Right $ Just $ RespTxOutputSpendStatus txss
                 _____ -> return $ RPCResponse 400 pretty $ Left $ RPCError INVALID_PARAMS Nothing
         _____ -> return $ RPCResponse 400 pretty $ Left $ RPCError INVALID_METHOD Nothing
-
-convertToScriptHash :: Network -> String -> Maybe String
-convertToScriptHash net s = do
-    let addr = stringToAddr net (DT.pack s)
-    (DT.unpack . txHashToHex . TxHash . sha256 . addressToScriptBS) <$> addr
-
-getNextCursor :: [ResultWithCursor r c] -> Maybe c
-getNextCursor [] = Nothing
-getNextCursor aos =
-    let nextCursor = cur $ last aos
-     in Just nextCursor
-
--- encode/decode NominalTxIndex and Output cursor types
-encodeNTI :: Maybe Int64 -> Maybe String
-encodeNTI mbNTI = show <$> mbNTI
-
-decodeNTI :: Maybe String -> Maybe Int64
-decodeNTI Nothing = Nothing
-decodeNTI (Just nti) = readMaybe nti :: Maybe Int64
-
-encodeOP :: Maybe (DT.Text, Int32) -> Maybe String
-encodeOP Nothing = Nothing
-encodeOP (Just op) = Just $ (DT.unpack $ fst op) ++ (show $ snd op)
-
-decodeOP :: Maybe String -> Maybe (DT.Text, Int32)
-decodeOP Nothing = Nothing
-decodeOP (Just c)
-    | length c < 65 = Nothing
-    | otherwise =
-        case readMaybe mbIndex :: Maybe Int32 of
-            Nothing -> Nothing
-            Just index -> Just (DT.pack txid, index)
-  where
-    (txid, mbIndex) = L.splitAt 64 c
